Enable PostCSS processing of custom CSS media and properties
The @custom-media rule is very much still experimental, and not
supported in browsers, but it's useful to be able to define some
constants. Expecting to have a small number of "breakpoints" that are
consistent across the whole site rather than ad-hoc rules in each
component/page.

As for custom properties, we're so far not planning to use them outside
a plain :root declaration block, so might as well process them as well.