Rename PageContainer to GridWrapper
Was repurposing this component in a way that didn't really fit so well
with the name. Don't expect to need a container for the page itself, but
will need containers for individual sections etc, because a section
might have a different background than the page itself, and that
background should fill the entire screen even though the layout grid
generally doesn't.

Can either accomplish this by a wrapper component, like this one, or by
repeating three CSS declarations everywhere the wrapper is needed (which
seems slightly too much).