Make main header sticky on all pages
Fixed-positioned on mobile (the site footer will have to be given enough
padding to ensure that all page contents can be viewed without a menu
button overlapping, given enough scrolling). Sticky on wider screens.

Sticky positioning is not supported by IE11, but supporting that browser
is not required for this project. Regardless, it should fall back
gracefully to the default 'position: static' styling.