Avoid className "mixin" in footer component
Would prefer to remove this ability from Paragraph, so it is "atomic" in
such a sense that it's immutable (as far as this is possible with CSS).

Having one element be affected by (non-"base") style rules from multiple
disjoint sources, rather than being "encapsulated" within a single
component (or page) that "owns" all responsibility for it, tends to
turn out to be unwieldy w.r.t. maintainability. (E.g. potential
conflicts would behave unpredictably, because we don't know which order
the declarations appear in the final combined stylesheet after CSS
Modules processing.)