Add styling to RichText
This currently has more variations in the Figma sketches than the Sanity
content model really contains, so these styles are picked from the
yellow boxes on the front page sketch combined with the headings on the
About page, to try to make some sort of temporary generic solution.
Perhaps a non-yellow background (without the vertical padding) would be
a better default, but this way we'll see the more "complex" variation.