Use images instead of Unicode for icons in Tickets component
Primarily, this is to be able to specify text alternatives. The "HEAVY
CHECK MARK" codepoint when used to represent "Included" might be
understandable in context, but "BALLOT X" for "Not included" does not
seem clear.

Secondarily, this aims to achieve a more consistent rendering. The
original Unicode codepoints are not supported by our web font, so we
were relying on some system font to handle it. (Worst case, it could've
ended up as being rendered by a fallback font e.g. as a rectangle.)

Ignoring the Next warning because the Next image overall doesn't seem
very useful here. We can handle CLS via width/height attributes, and the
srcset stuff that Next/Image would add seems pointless for SVGs, just a
waste of bytes.