Strikethrough expired ticket prices instead of not-yet-available ones
The way the "Available from" field on a ticket price in Sanity is
actually used, it's effectively supposed to set the expiry for the
previous ticket price. So as discussed e.g. in
https://app.shortcut.com/sanity-io/story/15373/all-tickets-table-presentation-fix-pre-launch
we're not supposed to show a price as "crossed out" even if its
"Available from" time is yet to arrive. Instead it's crossed out if
there is a newer price point whose "Available from" has been reached.