Show Program days and times according to expected timezone
The expected timezone being the one that's shown in the heading, taken
from the program's (first) venue. For instance, if the venue is Oslo
(heading is "Oslo - Central European Summer Time"), the instant
"2022-05-25T04:00:00.000Z" is shown as "May 25" and "06:00".
If it's San Fransisco ("San Fransisco - Pacific Daylight Time"), the
same instant would be shown as "May 24" and "21:00".

This also fixes a warning about server-side content and client-side
"hydrated" content not matching, because the timestamps used to be
re-rendered in the client's timezone (e.g. when accessing the site from
Oslo, all the days/times would be formatted for Europe/Oslo).