Fix TypeScript errors in Programs
The Venue type was missing the new "timezone" field, which is now added.
Furthermore, Programs was calling getNonLocationTimezone with a string,
which its implementation also happens to support (it uses
date-fns-tz' formatInTimeZone which takes either a Date instance or an
ISO8601 string for the relevant parameter), but it was supposed to be a
Date per the type declaration.