Fix erroneous logo file update
Accidentally updated the logo of the _conference_ instead of the Sanity
logo in the footer. Had an expectation in advance that they would be
very similar. But of course as it turns out there actually is a visible
difference; mainly, the shaded areas of the letters are different. It's
subtle, but not _that_ subtle.