Replace HTML "→" with CSS "->" in CookieConsent
The Unicode RIGHTWARDS ARROW character is not supported by the web font,
and is inconsistent with the -> ligature used elsewhere.

Also, the arrow is just a stylistic detail, not part of the content
proper. Don't really want e.g. assistive technology to read out "Learn
more right arrow" or similar.