Use cache-busting URLs in rel=icon links
To do this, the links must be moved into _app because Next.js _document
doesn't support loading external resources this way.

Could also have kept it as-is (in public/static), but the downside seems
to be that those files seem to be set to "max-age=0" and
"must-revalidate", making the browser re-check for each page load
whether the icon is stale. (Meanwhile a downside of the cache-busting
approach is that the URLs are slightly longer.)