Apply "blank" setting on PortableText link marks
In other words, if an "External link" has been added to some rich text
and its "Open in new tab" switch has been enabled, then request that the
browser opens a new browsing context (tab) when the link is activated.
Else, leave it at the default.

It's apparently possible that the "Open in new tab" switch is neither
true nor false. In such a case, this commit currently defaults to false
(no new tab). This seems consistent with the advice that the Studio
links to for this switch (https://css-tricks.com/use-target_blank/ which
basically argues that the target should be left as default for most
cases). On the other hand, some parts of this site currently use _blank
in a somewhat uncritical manner (considering that "Just because you want
users to never leave your page" and "'Internal' and 'External' links are
different" are both called out as a "Bad Reason" in the css-tricks
article).