Open SimpleCallToAction links in new tab when specified
This is intended to add support for the new simpleCallToAction schema,
including 'blank' setting which is used to specify whether the link
should open in a new tab/window.

Only aims to support the cases where the old simpleCallToAction was
already working. There seem to be cases where internal links don't work;
planning to refactor the GROQ queries to fix such things in a separate
task/branch.