Specify GROQ queries for 'seo' and 'hero' more precisely
Factoring out query fragments for 'figure' and 'simpleCallToAction'
fields is intended to make it possible to update in one place if/when
the data model is modified or we need more fields (e.g. fetching
width/height for 'figure' images).

For now these queries are named in a pretty generic fashion, though in
principle we may need multiple variants in the future, e.g. specifying
one projection to be used for a session when it appears in a list and
another projection to be used when it appears in a more detailed view on
a separate page.