Use alternate solution for published/updated separator on article page
Using a particular pixel width to hide/show the vertical line character
is basically doomed to fail because the text lengths of the dates can
vary widely (e.g. "September 30" vs "May 1"). It's possible to use
background-image tricks to let the bar appear only if there isn't a line
break, but maybe simply using the existing breakpoints is good enough?