Delete broken 'from' date validation
After introducing the nested 'priceAndAvailability' structure in #119, this validation rule no longer works as intended. The rule's peramaters are only 'value', being the date string that's entered in the input, and 'context', being the entire document; this doesn't give me enough info to know which 'from' date is being edited. There might be a way to find the correct date in the correct 'priceAndAvialability' object belonging to the correct 'group', but how to do so is not obvious to me. I'm going to simple route of removing the validation, given the number of other things awaiting to be shipped. (cc @kmelve)