Fix font+font name of Tickets group name and price
The font variable names are based on the names in the Figma sketches
which basically follow the format "size name - weight name" where the "
- weight name" part is omitted for the "Regular"/400 weight. E.g. "UI/SM
- Semi" which "translates" into "--font-ui-sm-semi" with our
(undocumented) convention. (So given this context "medium-semibold" is
onsensical, because it's naming two different weights.)

Also, the variable names are intented to be ordered alphabetically, so
moving the declarations of these accordingly at the same time.

Finally, adding letter-spacing since the designer requested this
(although the visual differences seem extremely minute here).