[[...slug]].tsx: don't re-render on scrollTop update
Since we set 'scrolledFarEnough' directly in the event listener now, there's no real need to have the 'scrollTop' state hook anymore. Removing it fixes a considerable performance sink: previously, we would re-render all pages on every scrollTop change, since that's what calling a setXX hook does. Since we only need scrollTop to compute another value, we skip scrollTop as an intermediary and 'setScrolledFarEnough' instead, which is a boolean and causes only a re-render when it flips.