Use canonical article URLs in SimpleCallToAction
This is more consistent with internal links in block content.

Seem to have to jump through some hoops for TypeScript, if we simply
pass link.internal to getEntityPath here, it complains that an argument
with optional slug is not assignable to the 'entity' parameter, even
though there is an if check further up.