Make slug input to getEntityPath nullable
As pointed out in review of PR
https://github.com/sanity-io/structured-content-2022/pull/126
there is already a guard clause in the function body which explicitly
"safe-navigates" the slug, so it didn't really require a non-null
parameter. Making it nullable means we have to jump through fewer TS
hoops when passing the 'internal' object of a 'Link' type.