Default to empty string instead of '#' in getEntityPath
As per review comment in
https://github.com/sanity-io/structured-content-2022/pull/126, this
makes it less awkward for consumers of the function to provide a
fallback for (or otherwise handle) the cases where a "real" entity path
was not found.

Intuitively, it seems more correct to return null. This is not done
here, to limit the scope of the PR. (Passing null to next/link would
yield an error, but passing the empty string does not, so this should be
much "safer" to do without having to go through and update other places
that call the getEntityPath function.) Can't think of any legitimate
uses of the empty string anyway, an empty input field in Sanity seems to
result in the corresponding property being completely missing in the
JSON returned.