Reinstate feature checkmark alignment
This seems to have been lost when FeatureCheckmark was factored out of
Tickets. Affects e.g. centering in the Tickets table and the spacing
between list items in the FeatureSection lists.

This commit makes the component align itself, so the CSS declaration
doesn't need to be repeated. Delegating the alignment to the container
would be a legitimate alternative, though. In this case the container
could either use a wrapper element, or the className prop could be kept.
But in the latter case, the component should really not style itself,
because that would risk ending up with conflicting style rules,
affecting maintainability. So far we don't really use such
"style-delegating" components elsewhere, so the behavior would be a bit
inconsistent/surprising.

The spacing between icon and additional text in sponsorship tables will
be handled separately. (It was a bit too small anyway, compared to Figma
sketches.)