Fix heading usage in th in Sponsorships
This mirrors part of commit 951e83127bd4f4d25c4ee5b55225c02eb3ba98ba.
Headings are not valid inside <th>.

Also adjusting the heading level. In practice the sponsorships section
is likely to have a heading of its own (right now it seems to be
"Sponsorship Packages"), which would be an H2. So H3 seems like a better
default, indicating a sub-heading within the section. It's already like
that for Tickets.