Fix margin of price groups in Tickets
In Chromium, there is some default margin on dl. Need to override this
separately now that the <dl> is on the price group instead of the
overall price list (as it was before the grouping).

Not noticeable in Firefox, since it implements the suggested default
rendering in the HTML spec here. (No margin on a dl that is nested
inside a ul.)