Remove min-height from Card component
The Figma sketches (for "Article – Extended") actually have a lower
height for cards without image in them. Looks slightly akward if there
is a card without image alongside a card with image, but this might not
be an issue in practice. Could alternatively use 'vertical-align:
middle' if that's any better.