Honor prefers-reduced-motion for item-enter animations
Probably could have tried to be clever about the --item-enter-animation
property (setting it to 'none' if the media query triggers), but given
that variables.css is currently pre-processed, that ruleset would have
to go somewhere else. Simply wrapping the declaration blocks that
actually sets the animation property itself seems simpler and more
direct, making it more obvious that the preference is being considered.