Crop HighlightedSpeakerBlock image on mobile/tablet
As discussed in
https://app.shortcut.com/sanity-io/story/14187/speaker-details-front-end,
we're doing this instead of scaling the image non-uniformly (stretching
it horizontally) as was done previously.

Doesn't seem to be any clear way to ensure the image is cropped
correctly, presumably the "hot-spot" is no longer usable given that it
refers to a rectangle in the original image prior to server-side
scaling/cropping. Hence given that these are portraits of people, this
commit assumes that it's better to crop from the bottom than from the
top or both. (The example image used for testing would cut off the top
of the head in the latter case.)