Fix rounded corners for "HalfOval" shape
For NavBlock, this was simply accomplished by setting border-radius on
the "FakeItem" HTML element itself. Could have ported this over to
the new Shape component, but this detail is easy to miss (especially
because the shapes are now completely random), so the same problem could
arise again if yet another component is made that uses these SVGs. Seems
safer to update the SVG itself instead. And this seems to be possible
with clipPath/clip-path.