Update desktop min-height on speaker details page
Needed after ea1629dd0f791c8cac5bd60f8f2d256299229a87 because the
HighlightedSpeakerBlock is taller now. Don't want it to overflow.

Alternative fix would be to keep (something like) the old min-height and
set overflow-y:hidden, thereby allowing the shapes to be clipped but
still ensuring that the speaker image (if any) is not.