Handle missing speaker photo on speaker page
We can either just omit the photo entirely or replace it with an extra
shape. This commit does the latter (which also is presumably more in
sync with what we'll end up doing once this component is used for the
session details page as well, with either one or two images being
shown).