speakers/[slug].tsx: list Speaker's Sessions
Specifically, list the Sessions which are tied to a Program which in turn must be associated with a Venue. Grab said Venue's timezone, and calculate the 'start' time for the Session based on the Program's 'startDateTime' and the index of appearance in the Program's Sessions; add all the durations to the Sessions prior to the current one's to the Program's 'startDateTime' to calculate the offset.