Stop capitalizing speaker title/company on session page
There seems to be no indication in the Figma sketches that this is
supposed to be done, so presumably it was a side-effect of sharing the
same font as the role (which is capitalized because the Sanity strings
are "speaker" and "moderator").

(The only difference with current staging data is that the "of" in "Head
of Content Strategy Relations" used to be capitalized, unlike the Figma
sketch.)

It's worth noting that CSS 'text-transform: capitalize' will capitalize
_each_ word, not just the first, which is not necessarily what was
intended. So far, though, the roles consist of single words anyway.

The alternative of doing the transformation in JS (i.e. on the actual
content, not just the styled appearance) would not necessarily be
entirely ideal either, since there is no completely equivalent way do
accomplish it using the standard library. Something like
role[0].toLocaleUpperCase() + role.slice(1) would be close, though
(with some further modifications needed to support codepoints outside
the BMP).