FakeItem.tsx: only set background-color for non-shape FakeItem
The order of the previous rules seem to get mangled in production builds, so despite them being correctly ordered such that a shape will overwrite the background styles of a FakeItem if applied in development, the same doesn't happen in production. This commit removes the possiblity of this conflict happening by applying a background-color only to those items that don't have a shape SVG set.