Delete revalidate endpoint
As the site has grown in complexity, there's now rarely a 1:1 correlation between one entity and one page, with most pages instead querying and rendering multiple entities. The simplest way of ensuring the site is always up-to-date is to adopt sanity.io's approach of simply rebuilding the site when a change is picked up. A webhook has been set up in Studio which calls a Vercel webhook to rebuild and redeploy the site. This eliminates the need for this revalidate endpoint.