Adjust speaker borders on session details page
The :first-child selector makes no sense on the <a> elements, because
they're always the first child of their parent <li>. The result of this
move (in 6e89dd7bc7cbdf2f1be2d16c61c68448b59750e4) was that two adjacent
speakers would end up with two borders between them.

On the other hand, unlike the current speakers section, this commit
doesn't put the border on the list item because the intended behavior
for :hover turns out to be that the borders above the hovered item
should disappear/be overlapped.

Instead, a negative top margin should do the trick of avoiding double
borders (since the border color is currently fully opaque).

Furthermore, on :hover it's better to switch the border color to
transparent instead of removing the entire border. This avoids "jumpy"
behavior where the content reflows and moves up/down when a speaker link
is hovered.