Fix session page speaker links reflowing on hover
This used to happen on small screens because the links would receive
horizontal padding only when hovered. The result was a "jumpy" reflow as
the content box had less space than in the un-hovered state.

This commit moves the borders to the list items after all, contradicting
the previous commit. Keeping them on the link was awkward since we'd
then need another wrapper element to modify the background (since the
top/bottom borders would otherwise be affected by the border-radius).

Thus the end result is similar to the speakers section, except with some
extra negative margins to handle the behavior where hover background
obscures the borders.