Move heading out of Accordion component
This may be personal preference, but I find it harder to reason about
the components' styling when they aren't organized in a typical tree
fashion with each one being "atomic" markup-wise.

The accordion component is already more complex than this in that it
accepts arbitrary content to be inserted into each of its panels. But it
doesn't seem necessary to do the same for the heading.

With the alternative approach in this commit, we're not locked into the
accordion being one section with no content before or after the
accordion itself other than the heading. The parent component can decide
how to layout the accordion relative to the other contents, e.g.
whether/where any margins should be added.