Fix sectioning markup for Accordion
Previously, the headings were siblings to the h2 of the Programs
section. Semantically, they should really be sub-headings.

Now the headings are marked as h3 instead of h2, and a sectioning
element has been added to explicitly indicate the heading scope.

In theory it might be appropriate to make the heading level a prop to
the Accordion component so that it can be adjusted based on the nesting
level of the parent, but so far it doesn't seem likely that we'll need
other levels than h3.