Fix horizontal alignment of Programs+Tickets headings on small screens
Currently there is no max width for widths below --medium-up, so the new
H2 headings for these two components were offset to the right when
compared to other H2s on the page (e.g. RichText). Could be observed
e.g. with 700px viewport width.

This commit focuses on achieving consistency. It might be good to have a
max width for narrow ("mobile") viewports too, but such a change should
be done "globally".