Fix vertical margins of Tickets section
These were removed from the table when adding the heading (commit
6a5229ca3e1a6591990e5fcab67e3b2b46ff0168), but we still need them
"outside" the heading in the general case. This is not visible on
/registration-info at the moment, but that's only because the sections
above and below it still do have margins. The tickets section could
appear first or last in the document, however. In these cases margins
are needed to achieve the intended spacing between header/footer and
tickets section.

(The header/footer elements don't have margins themselves because we
don't have a guarantee that the design won't introduce some new
component that wants to be directly adjacent with no margins. Case in
point, the Hero component.)