Fix incomplete CTA implementation in Tickets section
The direct ButtonLink usage did not account for the "open in new tab"
setting, and could give a warning about href not matching on server vs
client (observed locally with an "internal"-type link). It also didn't
hook into the path util that yields the canonical URL for articles (and
handles other document types, if the link schema is updated in the
future to allow linking to those).

Using the existing SimpleCallToAction component should resolve all this.