Remove font-display:optional for Inter font
We're relying on the -> ligature to be used here, so we'd really prefer
the webfont to actually be used. With 'optional' the browser only waits
for a very short period for the font to load (with the text being
invisible in the meantime), and then just gives up and uses a fallback
font. This means we often see the "arrows" shown without a ligature.

This commit just removes the 'font-display' descriptor entirely, leaving
it to the browser to decide what strategy is most appropriate given its
circumstances.

Some alternatives are
- use 'swap', which is similar to 'optional' in that it very quickly
  shows a fallback font, but then actually swaps to the webfont when it
  loads
- use 'block', which shows "invisible" text for a longer period before
  it starts using a fallback, but also keeps waiting for the web font
- keep 'optional' (and thus quite often not end up using the Inter font
  family at all, probably ending up with only two weights and losing the
  medium/semibold details), and use (an) SVG(s) for the arrow instead

Not clear to me which one would be preferable, except the third seems
suboptimal. Hence giving the browser the power to be intelligent about
it.