Highlight speakers on session page with 1-2 speakers
Had originally planned to adapt and reuse HighlightedSpeakerBlock for
this, but there are so many small differences in the design between the
two, so it would get pretty messy. Still, the implementation is based on
a "fork" from that component.

The session page speakers also need to be shown with name and
title/company, to be implemented in a future commit.

Another possible future improvement would be to make the markup use one
single list, maybe even avoiding duplication of the 2nd speaker, instead
of having one div for each desktop column.