Rename constant for speakers section query
(or more accurately, the constant that defines the GROQ projection for
speakers in the speakers section)

It's not really specific to the front page, that's just where the
speakers section happens to be used at the moment (for featured
speakers). The section can appear in any 'page' document, specifically
we are now planning to use it in /speakers (to list all speakers, as
part of an index page).