Remove obsolete links from NavBlock
Now that the conference is over, we no longer want Tickets, Registration
or Sponsorship links.

If this repo/code is reused for other conferences in the future, it
would be preferable to define some generic logic for how to layout this
component, so that it can work with non-hardcoded links.